feat: add evaluator operator using gemini + firebase genkit
create an evaluator operator
that evaluates the a proposed agent $taskId written in natural language (eg. "exchange 1$ USDC for 500 $AIPEPE") using its LLM and only outputs 1 token:
Y - the request should be executed as is - it's good value (500 $AIPEPE is worth >= 1USD)
N - the request should be ignored ($AIPEPE is a scam, so it's worth zero)
any other response - propose an alternative action (500 $AIPEPE is only worth $0.5USD now, but shows good signal, based on this evaluator's opinion).

the app will provide a UI showing tasks as chat messages (in a chatroom with the $delegationManager address as room is, that has operators as participants).

an agent action has following structure:
- a human readable proposal (string) with proposed action
- the proposal might optionally have one or more:

@$chainId:$contractAddr - tagging a specific operator - will signify to all participants that the task is addressed at this specific operator, similar to a tag request in a group chat

metadata.json
{"parentTask": $taskId} // of the action that this action is a response (R)

$chainId:$contractAddr.$methodCall.json
// a payload of metadata describing a suggested transaction, that effectively executes the proposal